requirements.txt
Django>=2.2.9,<3.0.0
django-braces==1.13.0
django-handyhelpers==0.0.26
djangorestframework==3.11.0
Sphinx==1.8.5
